RequestHeaders() and AcceptMime() are better suited in \gp\tool\Headers
